migrating to new downsample version
